# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenQuyHoach API",
    "description": "OpenQuyHoach — open, verifiable planning-data infrastructure for Vietnam.\n\nEvery entity carries provenance (source artifact sha256, retrieval time,\ntransformation steps, review state). Derived data is never presented as\nofficial.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/tiles/{z}/{x}/{y}.pbf": {
      "get": {
        "tags": [
          "publish"
        ],
        "summary": "Live Tile",
        "description": "Live MVT over published datasets. For snapshot tiles use the\npublication endpoints instead.",
        "operationId": "live_tile_v1_tiles__z___x___y__pbf_get",
        "parameters": [
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Z"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "X"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Y"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/publish": {
      "post": {
        "tags": [
          "publish"
        ],
        "summary": "Publish",
        "operationId": "publish_v1_publish_post",
        "parameters": [
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/publications": {
      "get": {
        "tags": [
          "publish"
        ],
        "summary": "List Publications",
        "operationId": "list_publications_v1_publications_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/publications/{pub_id}": {
      "get": {
        "tags": [
          "publish"
        ],
        "summary": "Get Publication",
        "operationId": "get_publication_v1_publications__pub_id__get",
        "parameters": [
          {
            "name": "pub_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pub Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/publications/{pub_id}/manifest.json": {
      "get": {
        "tags": [
          "publish"
        ],
        "summary": "Get Manifest",
        "operationId": "get_manifest_v1_publications__pub_id__manifest_json_get",
        "parameters": [
          {
            "name": "pub_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pub Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/publications/{pub_id}/tiles/{z}/{x}/{y}.pbf": {
      "get": {
        "tags": [
          "publish"
        ],
        "summary": "Publication Tile",
        "operationId": "publication_tile_v1_publications__pub_id__tiles__z___x___y__pbf_get",
        "parameters": [
          {
            "name": "pub_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pub Id"
            }
          },
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Z"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "X"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Y"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/compare": {
      "get": {
        "tags": [
          "publish"
        ],
        "summary": "Compare Layers",
        "operationId": "compare_layers_v1_compare_get",
        "parameters": [
          {
            "name": "from_layer",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "From Layer"
            }
          },
          {
            "name": "to_layer",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "To Layer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/changesets/{cs_id}": {
      "get": {
        "tags": [
          "publish"
        ],
        "summary": "Get Changeset",
        "operationId": "get_changeset_v1_changesets__cs_id__get",
        "parameters": [
          {
            "name": "cs_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cs Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/rasters/{dataset_id}/tiles/{z}/{x}/{y}.png": {
      "get": {
        "tags": [
          "publish"
        ],
        "summary": "Raster Tile",
        "description": "XYZ tile rendered from a published raster dataset's stored COG.",
        "operationId": "raster_tile_v1_rasters__dataset_id__tiles__z___x___y__png_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          },
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Z"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "X"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Y"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/versions/{version_id}/extent": {
      "get": {
        "tags": [
          "publish"
        ],
        "summary": "Version Extent",
        "description": "BBox + zoom hint for the map to frame a planning version.",
        "operationId": "version_extent_v1_versions__version_id__extent_get",
        "parameters": [
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublishRequest": {
        "properties": {
          "planning_version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Planning Version Id"
          },
          "max_zoom": {
            "type": "integer",
            "title": "Max Zoom",
            "default": 14
          },
          "allow_unreviewed": {
            "type": "boolean",
            "title": "Allow Unreviewed",
            "default": false
          },
          "async_": {
            "type": "boolean",
            "title": "Async",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "planning_version_id"
        ],
        "title": "PublishRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}